+++ NEWS.md
@@ -5,6 +5,10 @@ BUG FIXES
   - Setting `mean.path = FALSE` in `ggwithinstats` produced incorrect colors for
     points (#470). This bug was introduced in `0.6.5` and is now fixed.
 
+NEW FEATURES
+
+  - More models supported in `ggcoefstats`: `Glm`.
+
 # ggstatsplot 0.6.5
 
 BREAKING CHANGES